smolvlm/pyproject.toml
[project]
name = "smolvlm"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "transformers>=4.49.0,<5.0",
    "fastapi",
    "uvicorn[standard]",
    "pillow",
    "httpx",
    "python-multipart",
    "accelerate",
    "torchvision>=0.25.0",
    "num2words>=0.5.14",
]
